[XingfengYang] #LibC add adler32 hasher
