+++ SourceCode/Libraries/LibC/include/hash.h
@@ -1,0 +1,10 @@
+//
+// Created by XingfengYang on 2020/7/30.
+//
+#ifndef __LIBRARY_LIBC_HASH_H__
+#define __LIBRARY_LIBC_HASH_H__
+#include <stdint.h>
+
+uint32_t adler32(const char *buf, uint32_t buflength);
+
+#endif // __LIBRARY_LIBC_HASH_H__
